Issue #377 - use the VMUSB drivers in place rather than as copies
*  Fix error in the usb V1x90 code comparison was done with a = not
== corrupting the suffi8x.
*  Import support for the V1x90 from vmusb/daqconfig rather than a
local copy.
