+++ main/mvlc/Makefile.am
@@ -73,14 +73,14 @@ IMPORTED_MODULES=$(IMPORT_DIR)/CMQDC32RdoHdwr.h \
 	$(IMPORT_DIR)/CMDPP32SCP.cpp  $(IMPORT_DIR)/CMDPP32SCP.h \
 	$(IMPORT_DIR)/CMDPPQDC.h  \
 	$(IMPORT_DIR)/CMTDC32.h $(IMPORT_DIR)/CMTDC32.cpp \
-	$(IMPORT_DIR)/CNADC2530.h $(IMPORT_DIR)/CNADC2530.cpp
+	$(IMPORT_DIR)/CNADC2530.h $(IMPORT_DIR)/CNADC2530.cpp \
+	$(IMPORT_DIR)/CCAENV1x90Opcodes.h $(IMPORT_DIR)/CCAENV1x90Registers.h \
+	$(IMPORT_DIR)/CV1x90.cpp $(IMPORT_DIR)/CV1x90.h
 
 
 LOCAL_MODULES= 	MADC32Registers.h \
 	 CV977.h CV977.cpp \
 	CV1495sc.h CV1495sc.cpp CV1729.h CV1729.cpp \
-	CCAENV1x90Opcodes.h CCAENV1x90Registers.h CV1x90.cpp CV1x90.h \
-	sis3316.h sis3316Registers.h \
 	CXLM.h CXLM.cpp CXLMTimestamp.cpp CXLMTimestamp.h \
 	CXLMFERA.cpp CXLMFERA.h \
 	VMUSBCommand.h VMUSBCommand.cpp VMUSBListCommand.h \
